requirements.txt
# wheel
PyAstronomy
pyodbc==5.0.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.11
Django==5.0
django-appconf==1.0.6
django-compressor==4.4
django-filter==23.5
django-session-security==2.6.7
django-session-timeout==0.1.0
django-tailwind==3.7.0
django-thread==0.0.1
djlint==1.34.1
EditorConfig==0.12.3
et-xmlfile==1.1.0
future==0.18.3
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.6
jsbeautifier==1.14.11
json5==0.9.14
mysql-connector-python==8.2.0
mysqlclient==2.2.1
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
pathspec==0.12.1
pip==24.0
protobuf==4.21.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzk==0.9
rcssmin==1.1.1
regex==2023.10.3
requests==2.31.0
rjsmin==1.2.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.2.0
uwsgi==2.0.24
celery[redis]==5.3.5
redis==5.0.2
# channels==4.0.0
# flower==1.1.0